changes in the footer class.
Add some padding to it - add 40px to top and bottom and 100px on the 
left and right sides.

Add a 25px margin at the bottom of the footer. This will create a small 
space between the footer and the bottom part of the website.

Add a 30px margin at the bottom of the address tag.